+++ src/app/cart/cart.module.ts
@@ -1,20 +1,25 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { CartComponent } from './components/cart/cart.component';
+import { CartListComponent } from './components/cart-list/cart-list.component';
+import { CartItemComponent } from './components/cart-item/cart-item.component';
+
+import { SharedModule } from '../shared';
 
 import { CartService } from './services/cart.service';
 
 
 @NgModule({
   declarations: [
-    CartComponent
+    CartListComponent,
+    CartItemComponent,
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    SharedModule
   ],
   exports: [
-    CartComponent
+    CartListComponent
   ],
   providers: [
     CartService