+++ src/app/cart/services/cart.service.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@angular/core';
+import { Subject } from 'rxjs';
 
 import { Product } from '../../products/models/product.model';
 
@@ -6,11 +7,31 @@ import { Product } from '../../products/models/product.model';
   providedIn: 'root'
 })
 export class CartService {
-  products: Product[] = [];
+  private onAddToCard = new Subject<Product>();
+  private productsInCart: Product[] = [];
+
+  onAddToCard$ = this.onAddToCard.asObservable();
+
+  get cartList() {
+    return this.productsInCart;
+  }
+
+  get count() {
+    return this.productsInCart.length;
+  }
+
+  get sum() {
+    return this.productsInCart.reduce((sun: number, curItem: Product) => sun += curItem.price, 0);
+  }
 
   constructor() { }
 
-  buy(product: Product): void {
-    this.products.push(product);
+  addToCart(product: Product): void {
+    this.productsInCart.push(product);
+    this.onAddToCard.next(product);
+  }
+
+  removeItem(itemIndex: number): void {
+    this.productsInCart.splice(itemIndex, 1);
   }
 }