+++ src/app/products/components/product-list/product-list.component.ts
@@ -3,6 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { Product } from '../../models/product.model';
 
 import { ProductService } from '../../services/product.service';
+import { CartService } from '../../../cart/services/cart.service';
 
 @Component({
   selector: 'app-product-list',
@@ -12,10 +13,14 @@ import { ProductService } from '../../services/product.service';
 export class ProductListComponent implements OnInit {
   products: Product[];
 
-  constructor(private productService: ProductService) { }
+  constructor(private productService: ProductService, private cartService: CartService) { }
 
   ngOnInit() {
     this.products = this.productService.getProducts();
+  }
+
+  addProductToCart(product: Product) {
+    this.cartService.addToCart(product);
   }
 
 }