+++ src/app/products/components/product/product.component.ts
@@ -1,27 +1,22 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, Input, Output, EventEmitter } from '@angular/core';
 
 import { Product, ProductCategory } from '../../models/product.model';
-
-import { CartService } from '../../../cart/services/cart.service';
 
 @Component({
   selector: 'app-product',
   templateUrl: './product.component.html',
   styleUrls: ['./product.component.css']
 })
-export class ProductComponent implements OnInit {
+export class ProductComponent {
 
   @Input() product: Product;
-  @Input() isPurchased = false;
 
-  constructor(private cartService: CartService) { }
+  @Output() addToCart = new EventEmitter<Product>();
 
-  ngOnInit() {
-  }
+  constructor() { }
 
-  onBuy(): void {
-    console.log('Product hase bought!');
-    this.cartService.buy(this.product);
+  onAddToCart(): void {
+    this.addToCart.emit(this.product);
   }
 
   getCategory(category: ProductCategory): string {