+++ src/app/products/index.js
@@ -1,4 +1,6 @@
-export { ProductComponent } from './components/product/product.component';
-export { ProductListComponent } from './components/product-list/product-list.component';
+export * from './components/product/product.component';
+export * from './components/product-list/product-list.component';
 
-export { Product, ProductCategory } from '../../models/product.model';
+export * from './models/product.model';
+
+export * from './products.module';