+++ src/app/shared/directives/bg-highlighter.directive.spec.ts
@@ -1,0 +1,8 @@
+import { BgHighlighterDirective } from './bg-highlighter.directive';
+
+describe('BgHighlighterDirective', () => {
+  it('should create an instance', () => {
+    const directive = new BgHighlighterDirective();
+    expect(directive).toBeTruthy();
+  });
+});