+++ src/app/shared/shared.module.ts
@@ -1,12 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
+import { BgHighlighterDirective } from './directives/bg-highlighter.directive';
 
 
 @NgModule({
-  declarations: [],
+  declarations: [
+    BgHighlighterDirective
+  ],
   imports: [
     CommonModule
+  ],
+  exports: [
+    BgHighlighterDirective
   ]
 })
 export class SharedModule { }